Adjoint source TF edited
Changed the time-frequency phase misfit adjoint source, such that it includes a bit more info in the output -- most notably, the value 'max_delay' which is used for many of my output plots. This value is the largest phase shift found within a window, and is generally quite indicative of what we visually see. The units are the same as the phase shift, so radians (so the name is a bit misleading, I know).