Changed plot_raydensity
Adapted the plot_raydensity function (in lasif/visualization.py), such that the ray density info can be returned out of the function. I did this in order to be able to store it to file (instead of recalculating it every time again). Also slightly changed the appearance of the output plot to my liking.